+++ Sources/SwiftFortuneWheel/Configuration/ImagePreferences.swift
@@ -20,40 +20,28 @@ public struct ImagePreferences {
     /// Prefered image size, required
     public var preferredSize: CGSize
     
-    /// Horizontal offset in slice from the center
-    public var horizontalOffset: CGFloat
+    /// Horizontal offset in slice from the center, default value is `0`
+    public var horizontalOffset: CGFloat = 0
     
     /// Vertical offset in slice from the center
     public var verticalOffset: CGFloat
     
-    /// Flip the text upside down
-    public var flipUpsideDown: Bool
+    /// Flip the text upside down, default value is `false`
+    public var flipUpsideDown: Bool = false
     
-    /// Background color
-    public var backgroundColor: SFWColor?
+    /// Background color, `optional`
+    public var backgroundColor: SFWColor? = nil
     
-    /// Tint color
-    public var tintColor: SFWColor?
+    /// Tint color, `optional`
+    public var tintColor: SFWColor? = nil
     
     /// Initiates a image preferences
     /// - Parameters:
     ///   - preferredSize: Prefered image size, required
     ///   - verticalOffset: Vertical offset in slice from the center, default value is `0`
-    ///   - horizontalOffset: Horizontal offset in slice from the center, default value is `0`
-    ///   - flipUpsideDown: Flip the text upside down, default value is `false`
-    ///   - tintColor: Tint color, `optional`
-    ///   - backgroundColor: Background color, `optional`
     public init(preferredSize: CGSize,
-                verticalOffset: CGFloat = 0,
-                horizontalOffset: CGFloat = 0,
-                flipUpsideDown: Bool = false,
-                tintColor: SFWColor? = nil,
-                backgroundColor: SFWColor? = nil) {
+                verticalOffset: CGFloat = 0) {
         self.preferredSize = preferredSize
-        self.horizontalOffset = horizontalOffset
         self.verticalOffset = verticalOffset
-        self.backgroundColor = backgroundColor
-        self.tintColor = tintColor
-        self.flipUpsideDown = flipUpsideDown
     }
 }